requirements/base.txt
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.0
coreschema==0.0.4
defusedxml==0.6.0
Django==2.2.1
django-allauth==0.39.1
django-cors-headers==2.0.2
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
idna==2.8
isort==4.3.20
itypes==1.1.0
Jinja2==2.10.1
Markdown==2.6.8
MarkupSafe==1.1.1
oauthlib==3.0.1
olefile==0.46
openapi-codec==1.3.2
Pillow==4.1.1
Pygments==2.2.0
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.1
requests==2.21.0
requests-oauthlib==1.2.0
simplejson==3.16.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.24.1
